college/requirements.txt
django
pymysql
mysqlclient
pillow
django_cleanup
openpyxl
djangorestframework
markdown
django-filter
psycopg2